[#75] refactor: repository, adapter 분리
