+++ live-chat-server/src/main/java/org/giglab/live/infrastructure/adapter/RoomBroadcastPublisher.java
@@ -1,10 +1,11 @@
-package org.giglab.live.infrastructure.redis.pubsub;
+package org.giglab.live.infrastructure.adapter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.giglab.live.application.port.persistence.BroadcastPort;
+import org.giglab.live.infrastructure.redis.pubsub.RedisPubSubChannel;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
