+++ live-chat-server/src/main/java/org/giglab/live/infrastructure/adapter/RoomInsightAdapter.java
@@ -1,17 +1,18 @@
-package org.giglab.live.infrastructure.mongo;
+package org.giglab.live.infrastructure.adapter;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.giglab.live.application.dto.stats.ChatStats;
 import org.giglab.live.application.dto.stats.ViewerStats;
 import org.giglab.live.application.port.persistence.RoomInsightPort;
+import org.giglab.live.infrastructure.mongo.MongoChatInsightRepository;
 import org.giglab.live.infrastructure.mongo.aggregator.ChatMessageAggregator;
 import org.giglab.live.infrastructure.mongo.aggregator.ViewerSessionAggregator;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class MongoRoomInsightAdapter implements RoomInsightPort {
+public class RoomInsightAdapter implements RoomInsightPort {
 
   private final ViewerSessionAggregator viewerSessionAggregator;
   private final ChatMessageAggregator chatMessageAggregator;
