+++ live-chat-server/src/main/java/org/giglab/live/infrastructure/adapter/ViewerSessionAdapter.java
@@ -1,4 +1,4 @@
-package org.giglab.live.infrastructure;
+package org.giglab.live.infrastructure.adapter;
 
 import java.time.Duration;
 import java.time.Instant;
